+++ mutant-cli/src/commands/ls.rs
@@ -1,7 +1,6 @@
 use humansize::{BINARY, format_size};
 use log::{debug, info};
-
-use mutant_lib::MutAnt;
+use mutant_lib::prelude::MutAnt;
 use std::process::ExitCode;
 
 pub async fn handle_ls(mutant: MutAnt, long: bool) -> ExitCode {